+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - EntityTablePanel.Config.SUMMARY_PANEL_VISIBLE configuration value added.
 - EntityPanel.editControlPanel() and editControlTablePanel() replaced with mainPanel(), related refactoring.
 - EntityEditComponentPanel.excludeComponentFromSelection() replaced with selectableComponents().
+- WindowDetailLayout now requests initial focus when the window is initialized.
 
 ## 0.17.32
 ### is.codion.common.core